% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{accept_best_solution}
\alias{accept_best_solution}
\title{Default acceptance function for optimizer (always accept the current score if it is smaller than the best one obtained before)}
\usage{
accept_best_solution(current_score, best_score, ...)
}
\arguments{
\item{current_score}{Score from the current optimizing iteration (scalar value, double)}

\item{best_score}{Score from the current optimizing iteration (scalar value, double)}

\item{...}{Ignored arguments that may be used by alternative acceptance functions}
}
\value{
Boolean, TRUE if current score should be taken as the new optimal score, FALSE otherwise
}
\description{
Default acceptance function for optimizer (always accept the current score if it is smaller than the best one obtained before)
}
\keyword{internal}
