% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_accept.R
\name{accept_leftmost_improvement}
\alias{accept_leftmost_improvement}
\title{Alternative acceptance function for optimizer. Accept current score if no )}
\usage{
accept_leftmost_improvement(current_score, best_score, tol = 1e-06, ...)
}
\arguments{
\item{current_score}{One- or multi-dimensional score from the current optimizing iteration (double or vector of doubles)}

\item{best_score}{Best one- or multi-dimensional score found so far (double or vector of doubles)}

\item{tol}{Tolerance value}

\item{...}{Ignored arguments that may be used by alternative acceptance functions}
}
\value{
Boolean, TRUE if current score should be taken as the new optimal score, FALSE otherwise
}
\description{
Alternative acceptance function for optimizer. Accept current score if no )
}
\keyword{internal}
