% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_accept.R
\name{accept_strict_improvement}
\alias{accept_strict_improvement}
\title{Default acceptance function. Accept current score if and only if all elements are less than or equal than in best score
and there's at least one improvement.}
\usage{
accept_strict_improvement(current_score, best_score, ...)
}
\arguments{
\item{current_score}{One- or multi-dimensional score from the current optimizing iteration (double or vector of doubles)}

\item{best_score}{Best one- or multi-dimensional score found so far (double or vector of doubles)}

\item{...}{Ignored arguments that may be used by alternative acceptance functions}
}
\value{
Boolean, TRUE if current score should be taken as the new optimal score, FALSE otherwise
}
\description{
Default acceptance function. Accept current score if and only if all elements are less than or equal than in best score
and there's at least one improvement.
}
\keyword{internal}
