% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_equal_df.R
\name{all_equal_df}
\alias{all_equal_df}
\title{Compare two data.frames.}
\usage{
all_equal_df(df1, df2)
}
\arguments{
\item{df1}{first \code{\link[=data.frame]{data.frame()}} to compare}

\item{df2}{second \code{data.frame()} to compare}
}
\value{
\code{TRUE} or \code{FALSE} in case differences are present
}
\description{
This will convert factors to characters and disregard
row and column order
}
\keyword{internal}
