% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{assign_score_optimize_shuffle}
\alias{assign_score_optimize_shuffle}
\title{Shuffles samples trying to improve the scoring function.}
\usage{
assign_score_optimize_shuffle(
  batch_container,
  samples = NULL,
  n_shuffle = 2,
  iterations = 1000
)
}
\arguments{
\item{batch_container}{An instance of \code{BatchContainer}.}

\item{samples}{A \code{data.frame} with sample information. Should be \code{NULL} if the \code{BatchContainer}
already has samples in it.}

\item{n_shuffle}{Number of elements to swap in every iteration. \emph{Note}: at the moment only 2 is
supported.}

\item{iterations}{Number of iterations.}
}
\value{
A matrix with scores. Every row is an iteration. The matrix size is
\code{c(iterations, 1 + length(bc$aux_scoring_f))}.
}
\description{
In every iteration shuffles several samples in the container. If the \code{bc$score()} worsens,
reverts to the previous state.
}
