% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{countGini}
\alias{countGini}
\title{Gini index on counts}
\usage{
countGini(m)
}
\arguments{
\item{m}{a factor vector}
}
\value{
the value of the Gini index
}
\description{
Gini index on a factor vector of counts using the Gini function
from the package ineq
}
\author{
Juliane Siebourg-Polster
}
