% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{distribute_random}
\alias{distribute_random}
\title{Dummy distribution function which distributes samples randomly.}
\usage{
distribute_random(samples, batch_container)
}
\arguments{
\item{samples}{data.frame with samples and unique .sample_id field.}

\item{batch_container}{Instance of BatchContainer class}
}
\description{
Dummy distribution function which distributes samples randomly.
}
