% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{distribute_samples}
\alias{distribute_samples}
\title{General function for performing sample assignment.}
\usage{
distribute_samples(samples, batch_container,
  distribution_function = distribute_random, random_seed = NULL)
}
\arguments{
\item{samples}{data.frame with samples and unique .sample_id field.}

\item{batch_container}{Instance of BatchContainer class}

\item{random_seed}{If set will fix random seed and then return the original value}

\item{sitribution_function}{Function performing sample distribution}
}
\description{
General function for performing sample assignment.
}
