% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_cache.R
\name{.load_cache}
\alias{.load_cache}
\title{Load cache for a vignette chunk if it exists and
\code{params$regenerate_cache} is \code{FALSE}.}
\usage{
.load_cache(id)
}
\arguments{
\item{id}{A unique cache id.}
}
\description{
This function will set the following variables in the global environment:
\code{.cache_filename_write}, \code{.eval_cached_chunk}.
}
\details{
\code{.eval_cached_chunk} is set to \code{TRUE} if the chunk needs reevaluation.
}
\keyword{internal}
