% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_cache.R
\name{.save_cache}
\alias{.save_cache}
\title{Saves passed variables to \code{.cache_filename_write} if
\code{.cache_filename_write} is not \code{NULL} and \code{.eval_cached_chunk}
is \code{TRUE}}
\usage{
.save_cache(...)
}
\description{
Saves passed variables to \code{.cache_filename_write} if
\code{.cache_filename_write} is not \code{NULL} and \code{.eval_cached_chunk}
is \code{TRUE}
}
\keyword{internal}
