% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{drop_order}
\alias{drop_order}
\title{Drop highest order interactions}
\usage{
drop_order(.terms, m = -1)
}
\arguments{
\item{.terms}{\code{\link[stats:terms.object]{terms.object}}}

\item{m}{order of interaction (highest available if -1)}
}
\description{
Drop highest order interactions
}
