% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_blocks.R
\name{generate_blocks}
\alias{generate_blocks}
\title{Generate block design}
\usage{
generate_blocks(.tbl, k, ..., .name = "Batch", .prefix = "B",
  .control = list())
}
\arguments{
\item{.tbl}{a \code{\link[tibble:tibble]{tibble}} (or \code{data.frame}).}

\item{k}{number of batches (blocks), a single integer, a vector of batch lengths, or a vector of batch IDs.}

\item{...}{columns to select (see \code{\link[dplyr:select]{dplyr::select()}}).}

\item{.name}{name of the batch (block) column.}

\item{.prefix}{prefix of batch values ("B" for B1, B2, etc.).}

\item{.control}{list o}
}
\value{
sample list with the batch
}
\description{
Generate block design
}
