% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{generate_terms}
\alias{generate_terms}
\title{Generate \code{terms.object} (formula with attributes)}
\usage{
generate_terms(.tbl, ...)
}
\arguments{
\item{.tbl}{data}

\item{...}{columns to skip (unquoted)}
}
\value{
\code{\link[stats:terms.object]{terms.object}}
}
\description{
Generate \code{terms.object} (formula with attributes)
}
