% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{get_order}
\alias{get_order}
\title{Get highest order interaction}
\usage{
get_order(.terms)
}
\arguments{
\item{.terms}{\code{\link[stats:terms.object]{terms.object}}}
}
\value{
highest order (numeric).
}
\description{
Get highest order interaction
}
