% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{getScore}
\alias{getScore}
\title{get Score}
\usage{
getScore(layout, balance, sc_groups, sc_tests, bal_weights = rep(1,
  length(balance)), sc_weights = rep(1, length(sc_groups)))
}
\arguments{
\item{layout}{a data.frame with the factors to be balanced and their current positions}

\item{balance}{a vector with the names of experimental conditions to be balanced}

\item{sc_groups}{list of dimension name groups to be used jointly for scoring}

\item{sc_tests}{list of tests to use for each scoring group}

\item{bal_weights}{named vector of weights for the factors to be balanced (default all 1)}

\item{sc_weights}{named vector of weights for the dimensions (default all 1)}
}
\value{
the summarized penalty score
}
\description{
Summary score for an experimental layout based on the distribution of levels
of the factors to be balanced.
}
\author{
Juliane Siebourg-Polster
}
