% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invivo_samples}
\alias{invivo_samples}
\title{An artificial invivo sample list}
\format{
An object of class \code{"tibble"}
\describe{
\item{Study}{Study ID}
\item{AnimalID}{The animal IDs}
\item{Sex}{Female (F) or Male (M)}
\item{Genotype}{Genotype (CAR - mutant, WT - wild type)}
\item{Bodyeight}{Weight of the anima (g)}
\item{Litter}{The litter IDs}
\item{Earmark}{Side (\emph{L} left, \emph{R} right, or \emph{B} both),
and quantity (1-3, unspecified if only 1) of earmarks}
\item{Litter_alloc}{Litter with which the animal can be allocated}
}
}
\usage{
data(invivo_samples)
}
\description{
An animal sample list with 7 columns.
}
\author{
Guido Steiner
}
\keyword{datasets}
