% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invivo_study_treatments}
\alias{invivo_study_treatments}
\title{A treatment list together with additional constraints on the strain and sex of animals}
\format{
An object of class \code{"tibble"}
\describe{
\item{Treatment}{The treatment to be given to an individual animal (1-3, plus a few untreated cases)}
\item{Strain}{Strain (A or B) - a constraint which kind of animal may receive the respective treatment}
\item{Sex}{Female (F) or Male (M) - a constraint which kind of animal may receive the respective treatment}
}
}
\usage{
data(invivo_study_treatments)
}
\description{
This treatment list is intended to be used in connection with the \code{"invivo_study_samples"} data object
}
\author{
Guido Steiner
}
\keyword{datasets}
