% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{kruskal}
\alias{kruskal}
\title{Kruskal Wallis on Layout dimension}
\usage{
kruskal(m)
}
\arguments{
\item{m}{a factor vector}
}
\value{
the Kruskal-Wallis rank sum statistic
}
\description{
Kruskal-Wallis test statistic from the ranking of a factor vector
TODO: This is currently not used, it does not optimize for the correct layout
}
\author{
Juliane Siebourg-Polster
}
