% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_aggregation.R
\name{L1_norm}
\alias{L1_norm}
\title{Aggregation of scores: L1 norm}
\usage{
L1_norm(scores, ...)
}
\arguments{
\item{scores}{A score or multiple component score vector}

\item{...}{Parameters to be ignored by this aggregation function}
}
\value{
The L1 norm as an aggregated score.
}
\description{
This function enables comparison of the results of two scoring functions by calculating
an L1 norm (Manhattan distance from origin).
}
\examples{
L1_norm(c(2, 2))
}
