% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_aggregation.R
\name{L2s_norm}
\alias{L2s_norm}
\title{Aggregation of scores: L2 norm squared}
\usage{
L2s_norm(scores, ...)
}
\arguments{
\item{scores}{A score or multiple component score vector}

\item{...}{Parameters to be ignored by this aggregation function}
}
\value{
The squared L2 norm as an aggregated score.
}
\description{
This function enables comparison of the results of two scoring functions by calculating
an L2 norm (euclidean distance from origin). Since this is only used for ranking solutions,
the squared L2 norm is returned.
}
\examples{
L2s_norm(c(2, 2))
}
