% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_container.R
\name{locations_table_from_dimensions}
\alias{locations_table_from_dimensions}
\title{Create locations table from dimensions and exclude table}
\usage{
locations_table_from_dimensions(dimensions, exclude)
}
\arguments{
\item{dimensions}{A vector or list of dimensions. Every dimension
should have a name. Could be an integer vector of dimensions or
a named list. Every value of a list could be either dimension size
or parameters for \link[=BatchContainerDimension]{BatchContainerDimension$new()}.}

\item{exclude}{\link{data.frame} with excluded locations of a container.}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} with all the available locations.
}
\description{
Create locations table from dimensions and exclude table
}
