% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{make_colnames}
\alias{make_colnames}
\title{Make \link{matrix} column names unique.}
\usage{
make_colnames(m, prefix = "X")
}
\arguments{
\item{prefix}{Prefix to add if column names are empty.}
}
\value{
A \link{matrix} with updated column names.
}
\description{
Make \link{matrix} column names unique.
}
\keyword{internal}
