% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{meanDiff}
\alias{meanDiff}
\title{Mean difference}
\usage{
meanDiff(m)
}
\arguments{
\item{m}{a numeric vector}
}
\value{
the mean difference
}
\description{
Mean differens on vector to score continuous variables
}
\author{
Juliane Siebourg-Polster
}
