% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_accept.R
\name{mk_accept_leftmost_improvement_with_tolerance}
\alias{mk_accept_leftmost_improvement_with_tolerance}
\title{Create acceptance function in which order of scores (first to last) denotes relevance, using a tolerance parameter as well}
\usage{
mk_accept_leftmost_improvement_with_tolerance(tolerance = 1e-06)
}
\arguments{
\item{tolerance}{Tolerance value: When comparing score vectors from left to right, differences within +/- tol won't immediately
shortcut the comparison at this point, allowing improvement in a less important score to exhibit some influence}
}
\value{
Acceptance function accept_leftmost_improvement, but using a fixed tolerance parameter
}
\description{
Create acceptance function in which order of scores (first to last) denotes relevance, using a tolerance parameter as well
}
