% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{mk_complete_random_shuffling_function}
\alias{mk_complete_random_shuffling_function}
\title{Create function to reshuffle sample indices completely randomly}
\usage{
mk_complete_random_shuffling_function(n_samples)
}
\arguments{
\item{n_samples}{Total number of samples (i.e. max of permutation index)}
}
\value{
Parameter-less function to return a random permutation of the indices in the range 1:n_samples
}
\description{
This function was just added to test early on the functionality of optimize_design() to accept a permutation vector rather than a list with src and dst indices.
}
