% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{mk_pairwise_swapping_function}
\alias{mk_pairwise_swapping_function}
\title{Create function to propose 1 pairwise swap of samples on each call}
\usage{
mk_pairwise_swapping_function(n_samples)
}
\arguments{
\item{n_samples}{Total number of samples (i.e. max of permutation index)}
}
\value{
Parameter-less function to return a list with length 1 vectors 'src' and 'dst', denoting source and destination index for the swap operation
}
\description{
This internal function is wrapped by mk_swapping_function()
}
\keyword{internal}
