% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_annealing.R
\name{mk_simanneal_acceptance_func}
\alias{mk_simanneal_acceptance_func}
\title{Generate acceptance function for an optimization protocol based on simulated annealing}
\usage{
mk_simanneal_acceptance_func(
  temp_function = mk_simanneal_temp_func(T0 = 500, alpha = 0.8)
)
}
\arguments{
\item{temp_function}{A temperature function that returns the annealing temperature for a certain cycle k}
}
\value{
A function that takes parameters (\code{current_score}, \code{best_score}, \code{iteration}) for an optimization step and return a Boolean indicating whether the current solution should be accepted or dismissed. Acceptance probability of a worse solution decreases with annealing temperature.
}
\description{
Generate acceptance function for an optimization protocol based on simulated annealing
}
