% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_trt_day_samples}
\alias{multi_trt_day_samples}
\title{Unbalanced treatment and time sample list}
\format{
An object of class \code{"tibble"}
}
\usage{
data(multi_trt_day_samples)
}
\description{
A sample list with 4 columns SampleName, Well, Time and Treatment
Not all treatments are avaliable at all time points.
All samples are placed on the same plate.
}
\author{
siebourj
}
\keyword{datasets}
