% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{neighbors}
\alias{neighbors}
\title{Penalty for neighbors with same level}
\usage{
neighbors(m)
}
\arguments{
\item{m}{a factor vector}
}
\value{
the number of pairs
}
\description{
Penalty score for number of neighboring pairs with the same level of a factor.
The number of such pairs is summed.
}
\author{
Juliane Siebourg-Polster
}
