% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{OptimizationTrace}
\alias{OptimizationTrace}
\title{OptimizationTrace represents optimization trace.
Usually it is created by \code{\link[=optimize_design]{optimize_design()}}.}
\description{
OptimizationTrace represents optimization trace.
Usually it is created by \code{\link[=optimize_design]{optimize_design()}}.

OptimizationTrace represents optimization trace.
Usually it is created by \code{\link[=optimize_design]{optimize_design()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scores}}{Contains a matrix of scores. The matrix size is usually
\code{c(iterations + 1, length(bc$scoring_f))}}

\item{\code{aggregated_score}}{Contains a vector of aggregated scores. The vector length is usually
\code{c(iterations + 1)}}

\item{\code{seed}}{Saved value of \link{.Random.seed}.}

\item{\code{elapsed}}{Running time of the optimization.}

\item{\code{last_step}}{Last iteration step for which the score was set.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_steps}}{Returns number of steps in the \code{OptimizationTrace}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizationTrace$new()}}
\item \href{#method-set_scores}{\code{OptimizationTrace$set_scores()}}
\item \href{#method-shrink}{\code{OptimizationTrace$shrink()}}
\item \href{#method-get_scores}{\code{OptimizationTrace$get_scores()}}
\item \href{#method-print}{\code{OptimizationTrace$print()}}
\item \href{#method-plot}{\code{OptimizationTrace$plot()}}
\item \href{#method-clone}{\code{OptimizationTrace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{OptimizationTrace} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationTrace$new(n_steps, n_scores, score_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_steps}}{Number of values to save. Usually \code{n_steps == iterations + 1}.}

\item{\code{n_scores}}{Number of scoring functions.}

\item{\code{score_names}}{Names of scoring functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_scores"></a>}}
\if{latex}{\out{\hypertarget{method-set_scores}{}}}
\subsection{Method \code{set_scores()}}{
Set scores for i-th step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationTrace$set_scores(i, scores, aggregated_score)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Step number.}

\item{\code{scores}}{Scores, a vector or a value if no auxiliary functions are used.}

\item{\code{aggregated_score}}{Aggregated score, a double value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{OptimizationTrace} invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shrink"></a>}}
\if{latex}{\out{\hypertarget{method-shrink}{}}}
\subsection{Method \code{shrink()}}{
Shrink scores by keeping only first \code{last_step} scores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationTrace$shrink(last_step = self$last_step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{last_step}}{Last step to keep.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{OptimizationTrace} invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scores"></a>}}
\if{latex}{\out{\hypertarget{method-get_scores}{}}}
\subsection{Method \code{get_scores()}}{
Return individual (not aggregated!) scores by keeping only first \code{last_step} scores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationTrace$get_scores(last_step = self$last_step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{last_step}}{Last step to keep.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{OptimizationTrace} invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print \code{OptimizationTrace}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationTrace$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{OptimizationTrace} invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot \code{OptimizationTrace}. Only the main score at the moment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationTrace$plot(aggregated_only = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aggregated_only}}{only plot the aggregated score}

\item{\code{...}}{Extra arguments passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationTrace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
