% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osat.R
\name{osat_score_datatable}
\alias{osat_score_datatable}
\title{Compute OSAT score for sample assignment. This implementation is using datatable.}
\usage{
osat_score_datatable(df, batch_vars, feature_vars, expected_df = NULL)
}
\arguments{
\item{expected_df}{}
}
\value{
a list with two attributes: res$score (numberic score value), res$expected_df (expected counts dataframe for potential reuse)
}
\description{
Compute OSAT score for sample assignment. This implementation is using datatable.
}
\examples{
sample_assignment <- tibble::tribble(
  ~ID, ~SampleType, ~Race, ~plate,
  1, "Case", "Hispanic", 1,
  2, "Case", "Hispanic", 1,
  3, "Case", "European", 2,
  4, "Control", "Hispanic", 2,
  5, "Control", "European", 1,
  6, "Control", "European", 2,
  NA, NA, NA, 1,
  NA, NA, NA, 2,
)

osat_score_datatable(sample_assignment,
  batch_vars = "plate",
  feature_vars = c("SampleType", "Race")
)
}
