% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osat.R
\name{osat_score_tibble}
\alias{osat_score_tibble}
\title{Compute OSAT score for sample assignment. This implementation uses tibble.}
\usage{
osat_score_tibble(
  sample_assignment,
  batch_vars,
  feature_vars,
  expected_df = NULL
)
}
\arguments{
\item{sample_assignment}{\code{tibble} or \code{data.frame} where every row is a location
in a container and a sample in this location.}

\item{batch_vars}{vector with batch variables to take into account for the
score computation.}

\item{feature_vars}{vector with sample variables to take into account for
score comptutation.}

\item{expected_df}{A \code{tibble} with expected number of samples sample variables and
batch variables combination. This is not required, however it does not change during the
optimization proccess. So it is a good idea to cache this value.}
}
\value{
a list with two attributes: res$score (numeric score value), res$expected_df
(expected counts \code{tibble} for potential)
}
\description{
Compute OSAT score for sample assignment. This implementation uses tibble.
}
\examples{
sample_assignment <- tibble::tribble(
  ~ID, ~SampleType, ~Race, ~plate,
  1, "Case", "Hispanic", 1,
  2, "Case", "Hispanic", 1,
  3, "Case", "European", 2,
  4, "Control", "Hispanic", 2,
  5, "Control", "European", 1,
  6, "Control", "European", 2,
  NA, NA, NA, 1,
  NA, NA, NA, 2,
)

designit:::osat_score_tibble(sample_assignment,
  batch_vars = vars(plate),
  feature_vars = vars(SampleType, Race)
)
}
\keyword{internal}
