% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_samples.R
\name{pairwise_swapping}
\alias{pairwise_swapping}
\title{Proposes pairwise swap of samples on each call.}
\usage{
pairwise_swapping(batch_container, iteration)
}
\arguments{
\item{batch_container}{The batch-container.}

\item{iteration}{The current iteration number.}
}
\value{
Function accepting batch container & iteration number. It returns a list with length 1 vectors 'src' and 'dst', denoting source and destination index for the swap operation
}
\description{
This function will ensure that one of the locations is always non-empty. It should not
return trivial permutations (e.g., \code{src=c(1,2)} and \code{dst=c(1,2)}).
}
\keyword{internal}
