% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate_effect_example}
\alias{plate_effect_example}
\title{Example dataset with a plate effect}
\format{
An object of class \code{"tibble"}
\describe{
\item{row}{Plate row}
\item{column}{Plate column}
\item{conc}{Sample concentration}
\item{log_conc}{Logarithm of sample concentration}
\item{treatment}{Sample treatment}
\item{readout}{Readout from experiment}
}
}
\usage{
data(plate_effect_example)
}
\description{
Here top and bottom row were both used as controls (in dilutions). The top
row however was affected differently than the bottom one. This makes
normalization virtually impossible.
}
\author{
Balazs Banfai
}
\keyword{datasets}
