% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_design}
\alias{plot_design}
\title{Plot batch design}
\usage{
plot_design(.tbl, ..., .color, .alpha = NULL)
}
\arguments{
\item{.tbl}{a \code{\link[tibble:tibble]{tibble}} (or \code{data.frame}).}

\item{...}{columns to select for \code{x} axis (see \code{\link[dplyr:select]{dplyr::select()}}).}

\item{.color}{the variable used for coloring}

\item{.alpha}{the variable used for transparency}
}
\value{
the ggplot object
}
\description{
Plot batch design
}
