% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{random_score_variances}
\alias{random_score_variances}
\title{Estimate the variance of individual scores by a series of completely random sample permutations}
\usage{
random_score_variances(batch_container, random_perm, sample_attributes_fixed)
}
\arguments{
\item{batch_container}{An instance of \code{BatchContainer}.}

\item{random_perm}{Number of random sample permutations to be done.}

\item{sample_attributes_fixed}{Logical; if FALSE, simulate a shuffle function that alters sample attributes at each iteration.}
}
\value{
Vector of length m (=dimensionality of score) with estimated variances of each subscore
}
\description{
Estimate the variance of individual scores by a series of completely random sample permutations
}
\keyword{internal}
