% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{randomize}
\alias{randomize}
\title{randomize experimental layout}
\usage{
randomize(
  design,
  report,
  layout_dim,
  balance,
  scoring_groups = as.list(names(layout_dim)),
  scoring_tests = as.list(rep("countGini", length(layout_dim))),
  burnin = 100,
  annealprotocol,
  scoring_weights = rep(1, length(scoring_groups)),
  balance_weights = rep(1, length(balance)),
  distribute = 1:(prod(layout_dim))
)
}
\arguments{
\item{design}{a data.frame with the sample ids, experimental conditions and
information about fixed samples (columns with 'Fix' prefix and then the dimension name)}

\item{report}{a string with the sample identifier}

\item{layout_dim}{a named vector with the experimental dimensions}

\item{balance}{a vector with the names of experimental conditions to be balanced}

\item{scoring_groups}{list of dimension name groups to be used jointly for scoring}

\item{scoring_tests}{list of dimension name groups to be used jointly for scoring}

\item{burnin}{a number of initial burnin runs (default=100)}

\item{annealprotocol}{a vector with the number of pairs to swap in each annealing step}

\item{scoring_weights}{named vector of weights for the dimensions (default all 1)}

\item{balance_weights}{named vector of weights for the factors to be balanced (default all 1)}

\item{distribute}{a starting distribution}
}
\value{
the value of the Gini index
}
\description{
This function generates a randomized experimental layout based on given
experimental dimensions and factors to be balanced
}
\author{
Juliane Siebourg-Polster
}
