% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{report_scores}
\alias{report_scores}
\title{Helper function to print out one set of scores plus (if needed) aggregated values}
\usage{
report_scores(score, agg_score, iteration)
}
\arguments{
\item{score}{Vector of (non-aggregated) scores (may be a single value as well)}

\item{agg_score}{Vector of aggregated scores (may be a single value as well)}

\item{iteration}{Iteration number}
}
\description{
Helper function to print out one set of scores plus (if needed) aggregated values
}
\keyword{internal}
