% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_autoscaling.R
\name{sample_random_scores}
\alias{sample_random_scores}
\title{Sample scores from a number of completely random sample permutations}
\usage{
sample_random_scores(
  batch_container,
  scoring,
  random_perm,
  sample_attributes_fixed
)
}
\arguments{
\item{batch_container}{An instance of \link{BatchContainer}.}

\item{scoring}{A named \code{\link[=list]{list()}} of scoring function. Each function should
return a vector of non-zero length.}

\item{random_perm}{Number of random sample permutations to be done.}

\item{sample_attributes_fixed}{Logical; if \code{FALSE}, simulate a shuffle function that alters sample attributes at each iteration.}
}
\value{
A score matrix with n (# of permutations) rows and m (dimensionality of score) columns.
}
\description{
Sample scores from a number of completely random sample permutations
}
\keyword{internal}
