% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{shrink_mat}
\alias{shrink_mat}
\title{Shrinks a matrix with scores and adds an iteration column.}
\usage{
shrink_mat(m, last_iteration)
}
\arguments{
\item{m}{input matrix}

\item{last_iteration}{last iteration}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} wrapped in a list
}
\description{
Shrinks a matrix with scores and adds an iteration column.
}
\keyword{internal}
