% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_subgroups.R
\name{shuffle_with_subgroup_formation}
\alias{shuffle_with_subgroup_formation}
\title{Compose shuffling function based on already available subgrouping and allocation information}
\usage{
shuffle_with_subgroup_formation(
  subgroup_object,
  subgroup_allocations,
  keep_separate_vars = c(),
  report_grouping_as_attribute = FALSE
)
}
\arguments{
\item{subgroup_object}{A subgrouping object as returned by \code{form_homogeneous_subgroups()}}

\item{subgroup_allocations}{A list of possible assignments of the allocation variable as returned by \code{compile_possible_subgroup_allocation()}}

\item{keep_separate_vars}{Vector of column names in sample table; items with identical values in those variables will not be put into the same subgroup if at all possible}

\item{report_grouping_as_attribute}{Boolean, if TRUE, add an attribute table to the permutation functions' output, to be used in scoring during the design optimization}
}
\value{
Shuffling function that on each call returns an index vector for a valid sample permutation
}
\description{
Compose shuffling function based on already available subgrouping and allocation information
}
