% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_annealing.R
\name{simanneal_acceptance_prob}
\alias{simanneal_acceptance_prob}
\title{Acceptance probability for a new solution}
\usage{
simanneal_acceptance_prob(current_score, best_score, temp, eps = 0.1)
}
\arguments{
\item{current_score}{Score from the current optimizing iteration (scalar value, double)}

\item{best_score}{Score from the current optimizing iteration (scalar value, double)}

\item{temp}{Current value of annealing temperature}

\item{eps}{Small parameter eps(ilon), achieving that not always the new solution is taken when scores are exactly equal}
}
\value{
Probability with which to accept the new score as the best one
}
\description{
A solution is always to be accepted if it leads to a lower score. Worse solutions should be accepted with a probability given by this function.
}
\keyword{internal}
