% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_aggregation.R
\name{sum_scores}
\alias{sum_scores}
\title{Aggregation of scores: sum up all individual scores}
\usage{
sum_scores(scores, na.rm = FALSE, ...)
}
\arguments{
\item{scores}{A score or multiple component score vector}

\item{na.rm}{Boolean. Should NA values be ignored when obtaining the maximum? FALSE by default as ignoring NA values may render the sum meaningless.}

\item{...}{Parameters to be ignored by this aggregation function}
}
\value{
The aggregated score, i.e. the sum of all indicidual scores.
}
\description{
Aggregation of scores: sum up all individual scores
}
\examples{
sum_scores(c(3, 2, 1))
}
