% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{update_batchcontainer}
\alias{update_batchcontainer}
\title{Updates a batch container by permuting samples according to a shuffling}
\usage{
update_batchcontainer(bc, shuffle_params)
}
\arguments{
\item{bc}{The batch container to operate on.}

\item{shuffle_params}{Shuffling parameters as returned by \code{\link[=extract_shuffle_params]{extract_shuffle_params()}}.}
}
\value{
TRUE if sample attributes have been assigned, FALSE otherwise
}
\description{
As post-condition, the batch container is in a different state
}
\keyword{internal}
