% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_container.R
\name{validate_samples}
\alias{validate_samples}
\title{Validates sample data.frame.}
\usage{
validate_samples(samples)
}
\arguments{
\item{samples}{A \code{data.frame} having a sample annotation per row.}
}
\description{
Validates sample data.frame.
}
