% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_subgroups.R
\name{validate_subgrouping_object}
\alias{validate_subgrouping_object}
\title{Validate subgroup object and stop with error message if not all required fields are there}
\usage{
validate_subgrouping_object(subgroup_object)
}
\arguments{
\item{subgroup_object}{Subgrouping object as returned by \code{form_homogeneous_subgroups()}}
}
\description{
Validate subgroup object and stop with error message if not all required fields are there
}
\keyword{internal}
