% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{assign_random}
\alias{assign_random}
\title{Dummy assignment function which distributes samples randomly.}
\usage{
assign_random(samples, batch_container)
}
\arguments{
\item{samples}{data.frame with samples and unique .sample_id field.}

\item{batch_container}{Instance of BatchContainer class}
}
\description{
Dummy assignment function which distributes samples randomly.
}
