% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{assign_random}
\alias{assign_random}
\title{Dummy assignment function which distributes samples randomly.}
\usage{
assign_random(batch_container, samples = NULL)
}
\arguments{
\item{batch_container}{Instance of BatchContainer class}

\item{samples}{data.frame with samples and unique .sample_id field.}
}
\description{
Dummy assignment function which distributes samples randomly.
}
