% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{assign_score_optimize_shuffle}
\alias{assign_score_optimize_shuffle}
\title{Shuffles samples trying to improve the scoring function.}
\usage{
assign_score_optimize_shuffle(
  batch_container,
  samples = NULL,
  n_shuffle = NULL,
  shuffle_proposal = NULL,
  iterations = NULL,
  aggregate_scores_func = first_score_only
)
}
\arguments{
\item{batch_container}{An instance of \code{BatchContainer}.}

\item{samples}{A \code{data.frame} with sample information. Should be \code{NULL} if the \code{BatchContainer}
already has samples in it.}

\item{n_shuffle}{Number of times shuffling performed at each iterations.
Could be a number or a vector of length \code{iterations}.
In the later case, number of samples to shuffle could be precisely
set for every iteration. Please not that \code{n_shuffle} has a slightly different meaning depending
on whether you use \code{shuffle_proposal} or not. If \code{shuffle_proposal} is not set, this parameter
sets the number of items to shuffle (default 2). When \code{shuffle_proposal} is used, it will be
called \code{n_shuffle} times at an iteration (default 1).}

\item{shuffle_proposal}{A function used to propose two or more elements to shuffle in every step.
If non-\code{NULL} a function receives two arguments on every iteration:
\code{bc$samples_dt} and the iteration number. This function should return a list with attributes
\code{src} and \code{dst} (see \code{\link[=BatchContainer]{BatchContainer$exchange_samples()}}).}

\item{iterations}{Number of iterations. If not provided set to 1000.}

\item{aggregate_scores_func}{A function to aggregate the scores.
By default one is used that just uses the first score.}
}
\value{
An \link{OptimizationTrace} object.
}
\description{
In every iteration shuffles several samples in the container.
If the \code{batch_container$score()} worsens, reverts to the previous state.
}
