% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{assign_score_optimize_shuffle}
\alias{assign_score_optimize_shuffle}
\title{Shuffles samples trying to improve the scoring function.}
\usage{
assign_score_optimize_shuffle(
  batch_container,
  samples = NULL,
  n_shuffle = 1,
  shuffle_proposal = NULL,
  iterations = 1000
)
}
\arguments{
\item{batch_container}{An instance of \code{BatchContainer}.}

\item{samples}{A \code{data.frame} with sample information. Should be \code{NULL} if the \code{BatchContainer}
already has samples in it.}

\item{n_shuffle}{Number of times shuffling performed at each iterations.
Could be a number or a vector of length \code{iterations}.
In the later case, number of samples to shuffle could be precisely
set for every iteration. \code{n_shuffle == 1} indicates that two elements
are swapped one time. If \code{shuffle_proposal} function is provided, it will be
callend \code{n_shuffle} times at an iteration.}

\item{shuffle_proposal}{A function used to propose two or more elements to shuffle in every step.
If non-\code{NULL} a function receives two arguments on every iteration:
\code{bc$samples_dt} and the iteration number. This function should return a list with attributes
\code{src} and \code{dst} (see \code{\link[=BatchContainer$exchange_samples]{BatchContainer$exchange_samples()}}).}

\item{iterations}{Number of iterations.}
}
\value{
A matrix with scores. Every row is an iteration. The matrix size is
\code{c(iterations, 1 + length(bc$aux_scoring_f))}.
}
\description{
In every iteration shuffles several samples in the container.
If the \code{batch_container$score()} worsens, reverts to the previous state.
}
