% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{BatchContainer}
\alias{BatchContainer}
\title{R6 Class representing a batch container.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BatchContainer
}
\description{
R6 Class representing a batch container.
}
\keyword{datasets}
