% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{BatchContainerDimension}
\alias{BatchContainerDimension}
\title{R6 Class representing a batch container dimension.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BatchContainerDimension
}
\description{
R6 Class representing a batch container dimension.
}
\keyword{datasets}
