% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{BatchContainerDimension}
\alias{BatchContainerDimension}
\title{R6 Class representing a batch container dimension.}
\description{
R6 Class representing a batch container dimension.

R6 Class representing a batch container dimension.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{dimension name.}

\item{\code{values}}{vector of dimension values.}

\item{\code{weight}}{dimension weight. This can be interpreted
by the sample assignment function.}

\item{\code{parent_dimension}}{name of the parent dimension.
E.g., plate rows and columns belong to the plate.
However, there is no nesting between rows and columns.
This could be used by a sample assignment function
taking plate effect into account.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BatchContainerDimension$new()}}
\item \href{#method-clone}{\code{BatchContainerDimension$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchContainerDimension$new(
  name,
  size = NULL,
  values = NULL,
  weight = NA_real_,
  parent_dimension = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchContainerDimension$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
