% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cages}
\alias{cages}
\title{Mice sorted into cages}
\format{
A tibble with 40 rows and 15 variables:
\describe{
\item{Box}{Litter box}
\item{BoxSection}{Section of litter box}
\item{*Litter}{\code{Box} & \code{BoxSection}}
\item{*Sex}{Female (F) or Male (M)}
\item{Earmark}{Side (\emph{L} left, \emph{R} right, or \emph{B} both),
and quantity (1-3, unspecified if only 1) of earmarks}
\item{*AnimalID}{\code{Sex} - \code{Litter} - \code{Earmark}}
\item{*BirthDate}{Date of birth (date)}
\item{*Genotype}{Genotype (MUT - mutant, WT - wild type)}
\item{Comment}{Comments about the animal (two of them are deformed)}
\item{*ArrivalWeight}{Weight of the animal at arrival (g)}
\item{*Group}{\emph{(assigned)} Blinded treatment group (1 or 2)}
\item{CageNr}{\emph{(generated)} Cage number}
\item{CageCol}{\emph{(generated)} Column of cage rack}
\item{CageRow}{\emph{(generated)} Row of cage rack}
}
}
\usage{
cages
}
\description{
Original arrival data of 40 mice, their treatment assignments, and
their assignments into cages. Variables \code{Box}-\code{ArrivalWeight} are
based on the arrival data of the animals. \code{Group} variable is the
assigned treatment group, and \code{CageNr}, \code{CageCol}, and \code{CageRow}
variables are the generated cage layout.
Variables marked with an asterisk (*) are important for the cage layout
generation.
}
\author{
Guido Steiner
}
\keyword{data}
