% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cages}
\alias{cages}
\title{Mice sorted into cages}
\format{
A tibble with 40 rows and 15 variables:
\describe{
\item{Box}{Litter box}
\item{BoxSection}{Section of litter box}
\item{Litter}{Box&BoxSection}
\item{Sex}{Female (F) or Male (M)}
\item{Earmark}{Side (\emph{L} left, \emph{R} right, or \emph{B} both),
and quantity (1-3, unspecified if only 1) of earmarks}
\item{AnimalID}{}
\item{BirthDate}{Date of birth (date)}
\item{Genotype}{Genotype (MUT - mutant, WT - wild type)}
\item{Comment}{Comments about the animal}
\item{Gen}{SNPs of the animal}
\item{ArrivalWeight}{Weight of the animal at arrival (g)}
\item{Group}{Blinded treatment group (1 or 2)}
\item{CageNr}{Cage number}
\item{CageCol}{Column of cage rack}
\item{CageRow}{Row of cage rack}
}
}
\usage{
cages
}
\description{
Original arrival data of 40 mice, their treatment assignments, and
their assignments into cages.
}
\author{
Guido Steiner
}
\keyword{data}
