% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{complete_random_shuffling}
\alias{complete_random_shuffling}
\title{Reshuffle sample indices completely randomly}
\usage{
complete_random_shuffling(batch_container, iteration)
}
\arguments{
\item{batch_container}{The batch-container.}

\item{iteration}{The current iteration number.}
}
\value{
Parameter-less function to return a random permutation of the sample locations in the container
}
\description{
This function was just added to test early on the functionality of optimize_design() to accept a
permutation vector rather than a list with src and dst indices.
}
