% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designit-package.R
\docType{package}
\name{designit-package}
\alias{designit}
\alias{designit-package}
\title{designit: Blocking and randomization for experimental design}
\description{
Blocking and randomization for experimental design.
    TODO: expand this.
}
\author{
\strong{Maintainer}: Balazs Banfai \email{balazs.banfai@roche.com} (\href{https://orcid.org/0000-0003-0422-7977}{ORCID})

}
\keyword{internal}
