% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designit-package.R
\docType{package}
\name{designit-package}
\alias{designit}
\alias{designit-package}
\title{designit: Blocking and randomization for experimental design}
\description{
Blocking and randomization for experimental design.
    TODO: expand this.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.roche.com/#/designit/latest}
  \item \url{https://code.roche.com/BEDA/designit/}
  \item Report bugs at \url{https://code.roche.com/BEDA/designit/issues}
}

}
\author{
\strong{Maintainer}: Balazs Banfai \email{balazs.banfai@roche.com} (\href{https://orcid.org/0000-0003-0422-7977}{ORCID})

Authors:
\itemize{
  \item Juliane Siebourg-Polster \email{juliane.siebourg-polster@roche.com} (\href{https://orcid.org/0000-0002-1759-3223}{ORCID})
  \item Iakov Davydov \email{iakov.davydov@roche.com} (\href{https://orcid.org/0000-0003-3510-3926}{ORCID})
  \item Guido Steiner \email{guido.steiner@roche.com}
}

}
\keyword{internal}
