% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designit-package.R
\docType{package}
\name{designit-package}
\alias{designit}
\alias{designit-package}
\title{designit: Blocking and Randomization for Experimental Design}
\description{
Intelligently assign samples to batches in order to reduce batch effects. Batch effects can have a significant impact on data analysis, especially when the assignment of samples to batches coincides with the contrast groups being studied. By defining a batch container and a scoring function that reflects the contrasts, this package allows users to assign samples in a way that minimizes the potential impact of batch effects on the comparison of interest. Among other functionality, we provide an implementation for OSAT score by Yan et al. (2012, \doi{10.1186/1471-2164-13-689}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bedapub.github.io/designit/}
  \item \url{https://github.com/BEDApub/designit/}
  \item Report bugs at \url{https://github.com/BEDApub/designit/issues}
}

}
\author{
\strong{Maintainer}: Iakov I. Davydov \email{iakov.davydov@roche.com} (\href{https://orcid.org/0000-0003-3510-3926}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Juliane Siebourg-Polster \email{juliane.siebourg-polster@roche.com} (\href{https://orcid.org/0000-0002-1759-3223}{ORCID}) [copyright holder]
  \item Guido Steiner \email{guido.steiner@roche.com} [copyright holder]
  \item Jitao David Zhang \email{jitao_david.zhang@roche.com} (\href{https://orcid.org/0000-0002-3085-0909}{ORCID}) [copyright holder]
  \item Balazs Banfai \email{balazs.banfai@roche.com} (\href{https://orcid.org/0000-0003-0422-7977}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Konrad Rudolph \email{konrad.rudolph@roche.com} (\href{https://orcid.org/0000-0002-9866-7051}{ORCID}) [contributor]
  \item F. Hoffman-La Roche [copyright holder, funder]
}

}
\keyword{internal}
