% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datatable.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{This undocumented (?) flag allows us to use \verb{:=}
without getting an error.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
This undocumented (?) flag allows us to use \verb{:=}
without getting an error.
}
\examples{
\dontrun{
dt <- data.table::data.table(a=1, b=2)
dt[, a := NULL]
}
}
\keyword{internal}
