% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datatable.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{Show that the package is designed to rely on data.table functionality}
\description{
See \code{\link[data.table:assign]{data.table::let()}} and
\code{vignette("datatable-importing", "data.table")}
section "data.table in Imports but nothing imported".
}
\keyword{internal}
