% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{mk_autoscale_function}
\alias{mk_autoscale_function}
\title{Create a function that transforms a current (multi-dimensional) score into a boxcox normalized one}
\usage{
mk_autoscale_function(batch_container, random_perm)
}
\arguments{
\item{batch_container}{An instance of \code{BatchContainer}.}

\item{random_perm}{Number of random sample permutations for the estimation of the scaling params.}
}
\value{
The transformation function for a new score vector
}
\description{
Create a function that transforms a current (multi-dimensional) score into a boxcox normalized one
}
\keyword{internal}
