% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{mk_complete_random_shuffling_function}
\alias{mk_complete_random_shuffling_function}
\title{Create function to reshuffle sample indices completely randomly}
\usage{
mk_complete_random_shuffling_function(batch_container)
}
\arguments{
\item{batch_container}{A batch container with samples assigned that should be permuted randomly}
}
\value{
Parameter-less function to return a random permutation of the sample locations in the container
}
\description{
This function was just added to test early on the functionality of optimize_design() to accept a permutation vector rather than a list with src and dst indices.
}
