% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_plates.R
\name{mk_plate_scoring_functions}
\alias{mk_plate_scoring_functions}
\title{Create a list of scoring functions (one per plate) that quantify the spatially homogeneous distribution of conditions across the plate}
\usage{
mk_plate_scoring_functions(batch_container, plate = NULL, row, column, group)
}
\arguments{
\item{batch_container}{Batch container (bc) with all columns that denote plate related information}

\item{plate}{Name of the bc column that holds the plate identifier (may be missing or NULL in case just one plate is used)}

\item{row}{Name of the bc column that holds the plate row number (integer values starting at 1)}

\item{column}{Name of the bc column that holds the plate column number (integer values starting at 1)}

\item{group}{Name of the bc column that denotes a group/condition that should be distributed on the plate}
}
\value{
List of scoring functions, one per plate, that calculate a real valued measure for the quality of the group distribution (the lower the better)
}
\description{
Create a list of scoring functions (one per plate) that quantify the spatially homogeneous distribution of conditions across the plate
}
