% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{mk_subgroup_shuffling_function}
\alias{mk_subgroup_shuffling_function}
\title{Created a shuffling function that permutes samples within certain subgroups of the container locations}
\usage{
mk_subgroup_shuffling_function(
  subgroup_vars,
  restrain_on_subgroup_levels = c(),
  n_swaps = 1
)
}
\arguments{
\item{subgroup_vars}{Column names of the variables that together define the relevant subgroups}

\item{restrain_on_subgroup_levels}{Permutations can be forced to take place only within a level of the factor of the subgrouping variable. In this case, the user must pass only one subgrouping variable and a number of levels that together define the permuted subgroup.}

\item{n_swaps}{Vector with number of swaps to be proposed in successive calls to the returned function (each value should be in valid range from 1..floor(n_locations/2))}
}
\value{
Function to return a list with length n vectors 'src' and 'dst', denoting source and destination index for the swap operation, or NULL if the user provided a defined protocol for the number of swaps and the last iteration has been reached
}
\description{
If length(n_swaps)==1, the returned function may be called an arbitrary number of times.
If length(n_swaps)>1 the returned function may be called length(n_swaps) timed before returning NULL, which would be the stopping criterion if all requested swaps have been exhausted.
}
