% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osat.R
\name{osat_score}
\alias{osat_score}
\title{Compute OSAT score for sample assignment.}
\usage{
osat_score(df, batch_vars, feature_vars, expected_dt = NULL)
}
\arguments{
\item{batch_vars}{\code{character} vector with batch variable names to take into account for the
score computation.}

\item{feature_vars}{\code{character} vector with sample variable names to take into account for
score comptation.}

\item{expected_dt}{A \code{data.table} with expected number of samples sample variables and
batch variables combination. This is not required, however it does not change during the
optimization proccess. So it is a good idea to cache this value.}

\item{sample_assignment}{\code{data.table} or \code{data.frame} where every row is a location
in a container and a sample in this location.}
}
\value{
a list with two attributes: \code{$score} (numeric score value), \code{$expected_dt}
(expected counts \code{data.table} for reuse)
}
\description{
Compute OSAT score for sample assignment.
}
\examples{
sample_assignment <- tibble::tribble(
  ~ID, ~SampleType, ~Race, ~plate,
  1, "Case", "Hispanic", 1,
  2, "Case", "Hispanic", 1,
  3, "Case", "European", 2,
  4, "Control", "Hispanic", 2,
  5, "Control", "European", 1,
  6, "Control", "European", 2,
  NA, NA, NA, 1,
  NA, NA, NA, 2,
)

osat_score_datatable(sample_assignment,
  batch_vars = "plate",
  feature_vars = c("SampleType", "Race")
)
}
