% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osat.R
\name{osat_score_tibble}
\alias{osat_score_tibble}
\title{Compute OSAT score for sample assignment. This implementation uses tibble.}
\usage{
osat_score_tibble(
  sample_assignment,
  batch_vars,
  feature_vars,
  expected_df = NULL
)
}
\arguments{
\item{expected_df}{}
}
\value{
a list with two attributes: res$score (numberic score value), res$expected_df (expected counts dataframe for potential reuse)
}
\description{
Compute OSAT score for sample assignment. This implementation uses tibble.
}
\examples{
sample_assignment <- tibble::tribble(
  ~ID, ~SampleType, ~Race, ~plate,
  1, "Case", "Hispanic", 1,
  2, "Case", "Hispanic", 1,
  3, "Case", "European", 2,
  4, "Control", "Hispanic", 2,
  5, "Control", "European", 1,
  6, "Control", "European", 2,
  NA, NA, NA, 1,
  NA, NA, NA, 2,
)

osat_score(sample_assignment,
  batch_vars = vars(plate),
  feature_vars = vars(SampleType, Race)
)
}
