% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_plate}
\alias{plot_plate}
\title{Plot plate layouts}
\usage{
plot_plate(
  .tbl,
  Plate = Plate,
  Row = Row,
  Column = Column,
  .color,
  .alpha = NULL,
  .pattern = NULL
)
}
\arguments{
\item{.tbl}{a \code{\link[tibble:tibble]{tibble}} (or \code{data.frame}) with the samples assigned to locations}

\item{Plate}{the dimension variable used for the plate ids}

\item{Row}{the dimension variable used for the row ids}

\item{Column}{the dimension variable used for the column ids}

\item{.color}{the continuous or discrete variable to color by}

\item{.alpha}{a continuous variable encoding transparency}

\item{.pattern}{a discrete variable encoding tile pattern (needs ggpattern)}
}
\value{

}
\description{
Plot plate layouts
}
\examples{

nPlate <- 3
nColumn <- 4
nRow <- 6

treatments <- c("CTRL", "TRT1", "TRT2")
timepoints <- c(1, 2, 3)


bc <- BatchContainer$new(
  dimensions = list(
    plate = nPlate,
    column = list(values = letters[1:nColumn]),
    row = nRow
  )
)

sample_sheet <- tibble::tibble(
  sampleID = 1:(nPlate * nColumn * nRow),
  Treatment = rep(treatments, each = floor(nPlate * nColumn * nRow) / length(treatments)),
  Timepoint = rep(timepoints, floor(nPlate * nColumn * nRow) / length(treatments))
)

# assign samples from the sample sheet
assign_random(bc, samples = sample_sheet)

bc$get_samples()

plot_plate(bc$get_samples(),
  Plate = plate, Column = column, Row = row,
  .color = Treatment, .alpha = Timepoint
)

plot_plate(bc$get_samples(),
  Plate = plate, Column = column, Row = row,
  .color = Treatment, .pattern = Timepoint
)
}
\author{
siebourj
}
