% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{randomize}
\alias{randomize}
\title{randomize experimental layout}
\usage{
randomize(
  design,
  report,
  layoutDim,
  balance,
  factors = balance,
  scoringGroups = as.list(names(layoutDim)),
  scoringTests = as.list(rep("countGini", length(layoutDim))),
  burnin = 100,
  annealprotocol,
  scoringWeights = rep(1, length(scoringGroups)),
  balweights = rep(1, length(balance)),
  distribute = 1:(prod(layoutDim))
)
}
\arguments{
\item{design}{a data.frame with the sample ids, experimental conditions and
information about fixed samples (columns with 'Fix' prefix and then the dimension name)}

\item{report}{a string with the sample identifier}

\item{layoutDim}{a named vector with the experimental dimensions}

\item{balance}{a vector with the names of experimental conditions to be balanced}

\item{scoringGroups}{list of dimension name groups to be used jointly for scoring}

\item{scoringTests}{list of dimension name groups to be used jointly for scoring}

\item{burnin}{a number of initial burnin runs (default=100)}

\item{annealprotocol}{a vector with the number of pairs to swap in each annealing step}

\item{scoringWeights}{named vector of weights for the dimensions (default all 1)}

\item{balweights}{named vector of weights for the factors to be balanced (default all 1)}

\item{distribute}{a starting distribution}
}
\value{
the value of the Gini index
}
\description{
This function generates a randomized experimental layout based on given
experimental dimensions and factors to be balanced
}
\author{
Juliane Siebourg-Polster
}
